fair-service/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
psycopg2-binary
